entities hinzugefügt

+++ mrplauf/src/main/java/org/brandt/mrplauf/repositories/RessourceRepository.java
@@ -4,6 +4,6 @@ import org.brandt.mrplauf.entities.Produkt;
 import org.brandt.mrplauf.entities.Ressource;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RessourceRepository extends CrudRepository<Ressource, Long> {
+public interface RessourceRepository extends CrudRepository<Ressource, Integer> {
 
 }
